// FRONTEND/chatbot-admin-panel/src/services/api/schemas/documentSourceCreate.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * Mi Chatbot IA Personalizable
 * Un backend API en Python (FastAPI) para un chatbot avanzado con IA, configurable, capaz de conectarse a diversas fuentes de contexto y con gestión de permisos para administradores.
 * OpenAPI spec version: 0.2.0
 */
import type { DocumentSourceCreateDescription } from './documentSourceCreateDescription';
import type { DocSourceType } from './docSourceType';
import type { DocumentSourceCreatePathOrConfig } from './documentSourceCreatePathOrConfig';
import type { DocumentSourceCreateSyncFrequencyCron } from './documentSourceCreateSyncFrequencyCron';
import type { DocumentSourceCreateCredentialsInfo } from './documentSourceCreateCredentialsInfo';

export interface DocumentSourceCreate {
  /**
   * @minLength 3
   * @maxLength 100
   */
  name: string;
  description?: DocumentSourceCreateDescription;
  source_type: DocSourceType;
  /** Ruta (string) o configuración (JSON object) según el source_type. */
  path_or_config: DocumentSourceCreatePathOrConfig;
  is_active?: boolean;
  sync_frequency_cron?: DocumentSourceCreateSyncFrequencyCron;
  credentials_info?: DocumentSourceCreateCredentialsInfo;
}